old_projects/manageJobs_old/requirements.txt
Flask==3.0.0
flask-mongoengine==1.0.0
flask-login==0.6.3
flask-bcrypt==1.0.1
python-dotenv==1.0.0
gunicorn==21.2.0
email_validator==2.1.0
requests==2.31.0
werkzeug==3.0.1
mongoengine==0.27.0